+++ backend/canisters/local_user_index/impl/src/model/mod.rs
@@ -11,6 +11,7 @@ pub mod local_multi_user_map;
 pub mod local_user_map;
 pub mod media_scan_job_log;
 pub mod moderation_queue;
+pub mod multi_user_event_batch;
 pub mod premium_items;
 pub mod referral_codes;
 pub mod user_event_batch;