+++ backend/libraries/types/src/idempotency.rs
@@ -8,6 +8,16 @@ pub struct IdempotentEnvelope<T> {
     pub value: T,
 }
 
+impl<T> IdempotentEnvelope<T> {
+    pub fn map<U>(self, f: impl FnOnce(T) -> U) -> IdempotentEnvelope<U> {
+        IdempotentEnvelope {
+            created_at: self.created_at,
+            idempotency_id: self.idempotency_id,
+            value: f(self.value),
+        }
+    }
+}
+
 // Temp hack to allow us to release this in a non-breaking way
 impl<T> From<T> for IdempotentEnvelope<T> {
     fn from(value: T) -> Self {